+++ packages/metro-config/CHANGELOG.md
@@ -1,8 +1,16 @@
 # Change Log - @rnx-kit/metro-config
 
-This log was last generated on Tue, 09 Mar 2021 08:21:51 GMT and should not be manually modified.
+This log was last generated on Thu, 11 Mar 2021 15:31:28 GMT and should not be manually modified.
 
 <!-- Start content -->
+
+## 1.1.0
+
+Thu, 11 Mar 2021 15:31:28 GMT
+
+### Minor changes
+
+- Removes makeBabelConfig() for @rnx-kit/babel-preset-metro-react-native (4123478+tido64@users.noreply.github.com)
 
 ## 1.0.4